Move nats.aio.nuid to nats.nuid
