+++ nats/nuid.py
@@ -19,12 +19,11 @@ DIGITS = b'0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz'
 BASE = 62
 PREFIX_LENGTH = 12
 SEQ_LENGTH = 10
-TOTAL_LENGTH = PREFIX_LENGTH + SEQ_LENGTH
-MAX_SEQ = BASE**10
+MAX_SEQ = 839299365868340224 # BASE**10
 MIN_INC = 33
 MAX_INC = 333
 INC = MAX_INC - MIN_INC
-
+TOTAL_LENGTH = PREFIX_LENGTH + SEQ_LENGTH
 
 class NUID:
     """
@@ -39,21 +38,22 @@ class NUID:
         self._prefix = b''
         self.randomize_prefix()
 
-    def next(self):
+    def next(self) -> bytearray:
+        """
+        next returns the next unique identifier.
+        """
         self._seq += self._inc
         if self._seq >= MAX_SEQ:
             self.randomize_prefix()
             self.reset_sequential()
+
         l = self._seq
         prefix = self._prefix[:]
-
-        def _next():
-            nonlocal l
-            a = DIGITS[int(l) % BASE]
+        suffix = bytearray(SEQ_LENGTH)
+        for i in reversed(range(SEQ_LENGTH)):
+            suffix[i] = DIGITS[int(l) % BASE]
             l /= BASE
-            return a
 
-        suffix = bytearray(_next() for i in range(SEQ_LENGTH))
         prefix.extend(suffix)
         return prefix
